Update framework from jetpack.
Changes included:
*8dda03c: Adds coverage to GetByDocumentIdRequest.Builder
	 .addProjectionPaths
*c72b514: Not allowing joinspec in remove by query searchspecs
*3683fd2: Handle JoinSpec in ProtoConverter
*725c208: Adds an exception in SearchSpec builder
*3581e20: Adds @RequiresFeature to JoinSpec
*3f23348: Fix javadoc link and comment format
*cafd010: Moves DocumentIdUtilCtsTest to cts folder
*9dcfd97: Adds test coverage for SearchSpec.
*67806e9: Adds coverage to GetByDocumentIdRequest
*a4094b7: Adds a test to cover PropertyPath iterator()
*ad75d26: Adds a test for DocumentIdUtil.createQualifiedId with documents


Bug: 263152116
Bug: 256022027
Bug: 264468097
Bug: 261090345
Bug: 261090713
Bug: 261088722
Bug: 249829533
Test: presubmit

+++ framework/java/android/app/appsearch/AppSearchSession.java
@@ -692,6 +692,10 @@ public final class AppSearchSession implements Closeable {
         Objects.requireNonNull(executor);
         Objects.requireNonNull(callback);
         Preconditions.checkState(!mIsClosed, "AppSearchSession has already been closed");
+        if (searchSpec.getJoinSpec() != null) {
+            throw new IllegalArgumentException("JoinSpec not allowed in removeByQuery, but "
+                    + "JoinSpec was provided.");
+        }
         try {
             mService.removeByQuery(
                     mCallerAttributionSource,
